requirements.txt
Django==3.0.7
djangorestframework==3.11.0
pytest==5.4.3
pytest-django==3.9.0
